Separated the slider behavior classes
